[refactor] Login and Register screens
