+++ androidApp/src/main/java/tech/alexib/yaba/kmm/android/ui/plaid/PlaidLinkScreen.kt
@@ -13,7 +13,8 @@ import com.plaid.link.PlaidActivityResultContract
 import com.plaid.link.result.LinkResult
 import kotlinx.parcelize.Parcelize
 import org.koin.androidx.compose.getViewModel
-import tech.alexib.yaba.kmm.android.ui.defaultLogoBase64
+import tech.alexib.yaba.kmm.android.ui.components.defaultLogoBase64
+
 import tech.alexib.yaba.kmm.model.response.PlaidItemCreateResponse
